+++ dwertheimer.DateAutomations/src/index.js
@@ -5,4 +5,9 @@ export { insertDateTime8601 } from './dateFunctions'
 export { insertISODate } from './dateFunctions'
 export { insertCalendarNoteLink } from './dateFunctions'
 export { insertTime } from './dateFunctions'
+<<<<<<< HEAD
 export { dateFormatPicker } from './dateFunctions'
+=======
+export { datePicker } from './dateFunctions'
+export { insertStrftime } from './dateFunctions'
+>>>>>>> 2216198 (add formatted command)