+++ dwertheimer.DateAutomations/changelog.md
@@ -2,8 +2,12 @@ See Template Plugin [README](https://github.com/NotePlan/plugins/blob/main/dwert
 
 # What's Changed?
 
+<<<<<<< HEAD
 ## 1.0.9    
 - updated: now compiled for macOS versions back to 10.13.0.
+=======
+## 1.1.0    Added `formatted` option (using /formatted) and selecting `formatted` from `/dp` choice
+>>>>>>> 2216198 (add formatted command)
 
 ## 1.0.1    Removing "macOS.minVersion" - no longer necessary with transpiling